+++ eslint.config.js
@@ -1,0 +1,39 @@
+const { FlatCompat } = require('@eslint/eslintrc');
+const tsParser = require('@typescript-eslint/parser');
+const tsPlugin = require('@typescript-eslint/eslint-plugin');
+
+const compat = new FlatCompat({
+    baseDirectory: __dirname,
+    resolvePluginsRelativeTo: __dirname,
+});
+
+module.exports = [
+    {
+        files: ['src/**/*.ts'],
+        languageOptions: {
+            parser: tsParser,
+            parserOptions: {
+                ecmaVersion: 6,
+                sourceType: 'module',
+            },
+        },
+        plugins: {
+            '@typescript-eslint': tsPlugin,
+        },
+        rules: {
+            '@typescript-eslint/naming-convention': [
+                'warn',
+                {
+                    selector: 'import',
+                    format: ['camelCase', 'PascalCase'],
+                },
+            ],
+            '@/semi': 'warn',
+            curly: 'warn',
+            eqeqeq: 'warn',
+            'no-throw-literal': 'warn',
+            semi: 'off',
+        },
+        ignores: ['out', 'dist', '**/*.d.ts'],
+    },
+];