+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 
-All notable changes to the "fenced-div" extension will be documented in this file.
-
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.6.3] - 2025-01-15
+
+- Added support for exclamation marks !!!, MkDocs style
+
+## [0.5.3] - 2024-04-10
+
+- Refreshing preview to update after theme change on Mac
+
+## [0.5.1] - 2024-03-04
+
+- Fixed theme change
+
+## [0.5.0] - 2024-03-04
 
 - Initial release